feat(addon): add CreateFeatureAddonCommand and handler
- Implemented CreateFeatureAddonCommand to support creation of feature-based addons
- Added command handler to process addon creation logic
- Integrated validation and business rules for addon features
- Ensured repository and domain consistency with plan module

+++ Viora/Viora.Domain/Plans/Addons/LimitedFeatureAddon.cs
@@ -8,7 +8,7 @@ public class LimitedFeatureAddon : Entity
     public int RestoreValue { get; private set; }
     public double Price { get; private set; }
 
-    private LimitedFeatureAddon(Guid Id, Guid LimitedFeatureId, int RestoreValue, int price) : base(Id)
+    private LimitedFeatureAddon(Guid Id, Guid LimitedFeatureId, int RestoreValue, double price) : base(Id)
     {
         this.LimitedFeatureId = LimitedFeatureId;
         this.RestoreValue = RestoreValue;
